requirements.txt
tensorflow-gpu==2.0.0
numpy==1.18.1
nltk==3.4.5
scikit-learn
tqdm==4.46.0